Add keyboard shortcut ('cmd-shift-s') to entry in Plugins menu
This allows to invoke the plugin with a keyboard shortcut. Note that
many shortcuts are already used by Sketch and can't be used for plugins.
These include variations such as 'cmd-s', 'ctrl-s', 'cmd-alt-s'. If such
a shortcut is defined for a plugin, it is simply ignored and doesn't
show up in the Plugins menu (see [1]).

[1] https://developer.sketch.com/plugins/plugin-manifest#shortcut-values