Add scaling of style properties
Before only the frame [1] of each layer was scaled (which scales the
shape in itself), but not the style [2]. The style includes properties
such as the border width, shadow parameters, and text properties (font
size, line height, etc.).

Now, all the scalable property of the style of each layer are scaled
too. These properties include:

- Border width
- Shadow parameters (inner and outer shadow)
- Blur
- Text properties (font size, kerning, line height, paragraph spacing)

[1] https://developer.sketch.com/reference/api/#layer
[2] https://developer.sketch.com/reference/api/#style