Implement correct scaling of groups
Previously, groups [1] were scaled like other layers (i.e. scaling its
frame and style). However, for groups this didn't work because the style
of a group is mostly empty and the styles of the actual layers of a
group are defined in these layers themselves. That means, when a group
was scaled, the styles of the layers of the group weren't scaled (border
width, font size, etc.).

Now, groups are additionally scaled recursively by scaling each layer of
the group (which may again be a group) separately.

[1] https://developer.sketch.com/reference/api/#group